+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 ## New features
+- Added program models for `student_section_504_program_associations` and `student_neglected_or_delinquent_program_associations`
 ## Under the hood
 ## Fixes